Make auth manager indepentent
