+++ modules/ipfs-cpinner-client/src/constants.ts
@@ -5,5 +5,3 @@ export const SERVICE_MAX_STORAGE_REACHED = 'MAX_STORAGE_REACHED'
 export const MAX_STORAGE_REACHED = 'Max storage reached'
 export const AUTHENTICATION_ERROR = 'Authentication error'
 export const UNKNOWN_ERROR = 'Unknown error'
-export const NO_GET_ENCRYPTION_PUBLIC_KEY_FN = 'No getEncryptionPublicKey function present'
-export const NO_DECRYPT_FN = 'No decrypt function'
