+++ modules/ipfs-cpinner-client/src/encryption-manager/constants.ts
@@ -1,0 +1,3 @@
+// errors
+export const NO_GET_ENCRYPTION_PUBLIC_KEY_FN = 'No getEncryptionPublicKey function present'
+export const NO_DECRYPT_FN = 'No decrypt function'
