+++ modules/ipfs-cpinner-client/src/encryption-manager/types.ts
@@ -1,0 +1,12 @@
+export interface EncryptionManager {
+  encrypt(data: string): Promise<string>
+  decrypt(data: string): Promise<string>
+}
+
+export type GetEncryptionPublicKeyFn = () => Promise<string>
+export type DecryptFn = (data: string) => Promise<string>
+
+export type EncryptionManagerConfig = {
+  decrypt?: DecryptFn
+  getEncryptionPublicKey?: GetEncryptionPublicKeyFn
+}
