+++ modules/ipfs-cpinner-client/src/types.ts
@@ -1,4 +1,5 @@
 import { DIDAuthConfig } from './auth-manager/types'
+import { EncryptionManagerConfig } from './encryption-manager/types'
 
 export type GetContentPayload = { did: string, key: string }
 export type GetContentResponsePayload = { id: string, content: string }
@@ -9,15 +10,4 @@ export type SwapContentPayload = { key: string, content: string, id?: string }
 export type SwapContentResponse = { id: string }
 export type StorageInformation = { used: number, available: number }
 
-export interface EncryptionManager {
-  encrypt(data: string): Promise<string>
-  decrypt(data: string): Promise<string>
-}
-
-export type GetEncryptionPublicKeyFn = () => Promise<string>
-export type DecryptFn = (data: string) => Promise<string>
-
-export type Config = DIDAuthConfig & {
-  decrypt?: DecryptFn
-  getEncryptionPublicKey?: GetEncryptionPublicKeyFn
-}
+export type Config = DIDAuthConfig & EncryptionManagerConfig
