+++ modules/ipfs-cpinner-client/test/util.ts
@@ -9,7 +9,7 @@ import bodyParser from 'body-parser'
 import setupApi from '@rsksmart/ipfs-cpinner-service/lib/setup'
 import { Server } from 'http'
 import { hashPersonalMessage, ecsign, toRpcSig } from 'ethereumjs-util'
-import { DecryptFn, GetEncryptionPublicKeyFn } from '../src/types'
+import { DecryptFn, GetEncryptionPublicKeyFn } from '../src/encryption-manager/types'
 import { ClientKeyValueStorage } from '../src/auth-manager/types'
 import DataVaultWebClient from '../src'
 import authManagerFactory from '../src/auth-manager'
